Termino da primeira atividade

+++ exemplosDeCodigoJava/src/todasAtividadesDaCarreira/questaoCinco.java
@@ -8,12 +8,12 @@ public class questaoCinco {
         Scanner scan = new Scanner(System.in);
 
         System.out.println("Digite o valor do metro: ");
-        int valorMetro = scan.nextInt();
-        int valorCenti = 10 * 10;
-        int conversaoMetroParaCenti = valorMetro * valorCenti;
+        double valorMetro = scan.nextDouble();
+        double valorCenti = 10*10;
+        double conversaoMetroParaCenti = valorMetro * valorCenti;
 
         System.out.println("A conversão de metro para centímetros é de " + conversaoMetroParaCenti + " cm");
 
-
+        // Unico erro foi que a professora botou em double e não em int
     }
 }
